# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Packaging Compliance AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /packages:
    get:
      operationId: listPackages
      tags: [packages]
      summary: List packaging reviews with optional filters
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: risk
          in: query
          required: false
          schema:
            type: string
        - name: engine
          in: query
          required: false
          schema:
            type: string
        - name: vendor
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of packages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Package"
    post:
      operationId: createPackage
      tags: [packages]
      summary: Upload a package for review (auto-runs AI analysis when artwork text is provided)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PackageInput"
      responses:
        "201":
          description: Created package
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PackageDetail"
...
  /packages/{id}:
    get:
      operationId: getPackage
      tags: [packages]
      summary: Get a package with full compliance intelligence
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Package detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PackageDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    patch:
      operationId: updatePackage
      tags: [packages]
      summary: Update package status or workflow state
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PackageUpdate"
      responses:
        "200":
          description: Updated package
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PackageDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deletePackage
      tags: [packages]
      summary: Delete a package
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /packages/{id}/analyze:
    post:
      operationId: analyzePackage
      tags: [packages]
      summary: Run (or re-run) the AI compliance analysis pipeline
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Analyzed package detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PackageDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /packages/{id}/copilot:
    post:
      operationId: askCopilot
      tags: [packages]
      summary: Ask the AI compliance copilot a question about this package
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CopilotInput"
      responses:
        "200":
          description: Copilot answer with citations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CopilotAnswer"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /packages/{id}/audit:
    get:
      operationId: getPackageAudit
      tags: [packages]
      summary: Review history for a package
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Audit events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditEvent"
  /packages/{id}/report:
    post:
      operationId: generateReport
      tags: [reports]
      summary: Generate a compliance report for a package
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportInput"
      responses:
        "201":
          description: Created report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /packages/bulk-analyze:
    post:
      operationId: bulkAnalyze
      tags: [packages]
      summary: Run compliance analysis across many packages
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkAnalyzeInput"
      responses:
        "200":
          description: Bulk analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkAnalyzeResult"
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Bbox:
      type: object
      properties:
        x:
          type: number
        y:
          type: number
        w:
          type: number
        h:
          type: number
      required:
        - x
        - y
        - w
        - h
    Violation:
      type: object
      properties:
        id:
          type: integer
        packageId:
          type: integer
        severity:
          type: string
        engine:
          type: string
        title:
          type: string
        description:
          type: string
        regulationRef:
          type: ["string", "null"]
        recommendation:
          type: ["string", "null"]
        detectedText:
          type: ["string", "null"]
        suggestedText:
          type: ["string", "null"]
        bbox:
          oneOf:
            - $ref: "#/components/schemas/Bbox"
            - type: "null"
        status:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - packageId
        - severity
        - engine
        - title
        - description
        - status
        - createdAt
...
    OcrData:
      type: object
      properties:
        productName:
          type: ["string", "null"]
        ingredients:
          type: ["string", "null"]
        directions:
          type: ["string", "null"]
        warnings:
          type: ["string", "null"]
        claims:
          type: array
          items:
            type: string
        marketingCopy:
          type: ["string", "null"]
        nutritionFacts:
          type: ["string", "null"]
        allergenStatements:
          type: ["string", "null"]
        netWeight:
          type: ["string", "null"]
        countryOfOrigin:
          type: ["string", "null"]
        manufacturerInfo:
          type: ["string", "null"]
        expirationDate:
          type: ["string", "null"]
        epaRegistrationNumbers:
          type: ["string", "null"]
        hazardStatements:
          type: ["string", "null"]
    Package:
      type: object
      properties:
        id:
          type: integer
        sku:
          type: string
        upc:
          type: ["string", "null"]
        name:
          type: string
        brand:
          type: string
        vendor:
          type: string
        category:
          type: string
        country:
          type: ["string", "null"]
        netWeight:
          type: ["string", "null"]
        dimensions:
          type: ["string", "null"]
        packageType:
          type: ["string", "null"]
        productType:
          type: ["string", "null"]
        manufacturingRegion:
          type: ["string", "null"]
        status:
          type: string
        grade:
          type: ["string", "null"]
        riskScore:
          type: ["integer", "null"]
        complianceStatus:
          type: string
        reviewer:
          type: ["string", "null"]
        artworkUrl:
          type: ["string", "null"]
        summary:
          type: ["string", "null"]
        criticalCount:
          type: integer
        majorCount:
          type: integer
        minorCount:
          type: integer
        analyzedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - sku
        - name
        - brand
        - vendor
        - category
        - status
        - complianceStatus
        - criticalCount
        - majorCount
        - minorCount
        - createdAt
        - updatedAt
    PackageDetail:
      type: object
      properties:
        id:
          type: integer
        sku:
          type: string
        upc:
          type: ["string", "null"]
        name:
          type: string
        brand:
          type: string
        vendor:
          type: string
        category:
          type: string
        country:
          type: ["string", "null"]
        netWeight:
          type: ["string", "null"]
        dimensions:
          type: ["string", "null"]
        packageType:
          type: ["string", "null"]
        productType:
          type: ["string", "null"]
        manufacturingRegion:
          type: ["string", "null"]
        status:
          type: string
        grade:
          type: ["string", "null"]
        riskScore:
          type: ["integer", "null"]
        complianceStatus:
          type: string
        reviewer:
          type: ["string", "null"]
        artworkUrl:
          type: ["string", "null"]
        summary:
          type: ["string", "null"]
        extractedText:
          type: ["string", "null"]
        criticalCount:
          type: integer
        majorCount:
          type: integer
        minorCount:
          type: integer
        analyzedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        ocr:
          oneOf:
            - $ref: "#/components/schemas/OcrData"
            - type: "null"
        recommendations:
          type: array
          items:
            type: string
        violations:
          type: array
          items:
            $ref: "#/components/schemas/Violation"
        regulations:
          type: array
          items:
            $ref: "#/components/schemas/Regulation"
      required:
        - id
        - sku
        - name
        - brand
        - vendor
        - category
        - status
        - complianceStatus
        - criticalCount
        - majorCount
        - minorCount
        - createdAt
        - updatedAt
        - recommendations
        - violations
        - regulations
...
    PackageInput:
      type: object
      properties:
        sku:
          type: string
          minLength: 1
        upc:
          type: string
        name:
          type: string
          minLength: 1
        brand:
          type: string
          minLength: 1
        vendor:
          type: string
          minLength: 1
        category:
          type: string
        country:
          type: string
        netWeight:
          type: string
        dimensions:
          type: string
        packageType:
          type: string
        productType:
          type: string
        manufacturingRegion:
          type: string
        artworkUrl:
          type: string
        extractedText:
          type: string
      required:
        - sku
        - name
        - brand
        - vendor
    PackageUpdate:
      type: object
      properties:
        status:
          type: string
        reviewer:
          type: string
        grade:
          type: string
        riskScore:
          type: integer
        complianceStatus:
          type: string
    CopilotInput:
      type: object
      properties:
        question:
          type: string
          minLength: 1
      required:
        - question
    Citation:
      type: object
      properties:
        source:
          type: string
        section:
          type: ["string", "null"]
        text:
          type: ["string", "null"]
      required:
        - source
    CopilotAnswer:
      type: object
      properties:
        answer:
          type: string
        citations:
          type: array
          items:
            $ref: "#/components/schemas/Citation"
      required:
        - answer
        - citations
    BulkAnalyzeInput:
      type: object
      properties:
        ids:
          type: array
          items:
            type: integer
      required:
        - ids
    BulkAnalyzeResult:
      type: object
      properties:
        analyzed:
          type: integer
        passed:
          type: integer
        failed:
          type: integer
      required:
        - analyzed
        - passed
        - failed
    Regulation:
      type: object
      properties:
        id:
          type: integer
        agency:
          type: string
        category:
          type: string
        ruleCode:
          type: string
        title:
          type: string
        summary:
          type: string
        regulationText:
          type: ["string", "null"]
        section:
          type: ["string", "null"]
        source:
          type: ["string", "null"]
        publicationDate:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - agency
        - category
        - ruleCode
        - title
        - summary
        - createdAt
...
    AuditEvent:
      type: object
      properties:
        id:
          type: integer
        packageId:
          type: ["integer", "null"]
        entityType:
          type: ["string", "null"]
        entityId:
          type: ["integer", "null"]
        actor:
          type: string
        action:
          type: string
        detail:
          type: ["string", "null"]
        before:
          nullable: true
        after:
          nullable: true
        regulationRefs:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - actor
        - action
        - createdAt
...
    Report:
      type: object
      properties:
        id:
          type: integer
        packageId:
          type: ["integer", "null"]
        title:
          type: string
        type:
          type: string
        format:
          type: string
        summary:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - type
        - format
        - createdAt
    ReportInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        type:
          type: string
        format:
          type: string
      required:
        - title
...